+++ app/Controllers/Http/ProductController.js
@@ -2,8 +2,10 @@
 
 const productValidator = require("../../../service/productValidator");
 const Product = use("App/Models/Product");
+const Customer = use("App/Models/Customer");
 const makeProductUtil = require("../../../util/ProductUtil.func");
 const numberTypeParamValidator = require("../../../util/numberTypeParamValidator.func");
+const performAuthentication = require("../../../util/authenticate.func");
 
 class ProductController {
   async index({ request }) {
@@ -31,14 +33,41 @@ class ProductController {
     return { status: 200, error: undefined, data: product || {} };
   }
 
-  async store({ request }) {
+  async store({ auth, request }) {
     const { body, qs } = request;
 
-    const { customer_id, product_name, end_date, stock } = body;
+    const { product_name, end_date, stock } = body;
 
     const { references } = qs;
 
-    const validation = await productValidator(request.body);
+    const { admin, error } = await performAuthentication(auth).validateAdmin();
+
+    if (error)
+      return {
+        status: 403,
+        error,
+        data: undefined,
+      };
+
+    if (admin)
+      return {
+        status: 403,
+        error: "This action is reserved for regular user only.",
+        data: undefined,
+      };
+
+    const { customer_id } = await performAuthentication(auth).validateIdParam(
+      Customer
+    );
+
+    console.log(customer_id);
+
+    const validation = await productValidator({
+      customer_id,
+      product_name,
+      end_date,
+      stock,
+    });
 
     if (validation.error) {
       return { status: 422, error: validation.error, data: undefined };
@@ -61,33 +90,101 @@ class ProductController {
     };
   }
 
-  async update({ request }) {
+  async update({ auth, request }) {
     const { body, params, qs } = request;
 
     const { id } = params;
 
     const { references } = qs;
 
-    const { customer_id, product_name, end_date, stock } = body;
+    const { product_name, end_date, stock } = body;
 
-    const product = await makeProductUtil(Product).updateById(
-      id,
-      { customer_id, product_name, end_date, stock },
-      references
-    );
+    const { admin, error } = await performAuthentication(auth).validateAdmin();
 
-    return { status: 200, error: undefined, data: product };
+    if (error)
+      return {
+        status: 403,
+        error,
+        data: undefined,
+      };
+
+    const validateValue = numberTypeParamValidator(id);
+
+    if (validateValue.error)
+      return { status: 500, error: validateValue.error, date: undefined };
+
+    if (admin) {
+      const product = await makeProductUtil(Product).updateById(
+        id,
+        { product_name, end_date, stock },
+        references
+      );
+
+      return { status: 200, error: undefined, data: product };
+    }
+
+    const { customer_id } = await performAuthentication(auth).validateIdParam();
+
+    if (customer_id === parseInt(id)) {
+      const product = await makeProductUtil(Product).updateById(
+        customer_id,
+        { product_name, stock },
+        references
+      );
+
+      return { status: 200, error: undefined, data: product };
+    }
+
+    return {
+      status: 403,
+      error: "id param does not match credential id.",
+      data: undefined,
+    };
   }
 
-  async destroy({ request }) {
+  async destroy({ auth, request }) {
     const { id } = request.params;
 
-    const product = await makeProductUtil(Product).deleteById(id);
+    const { admin, error } = await performAuthentication(auth).validateAdmin();
+
+    if (error)
+      return {
+        status: 403,
+        error,
+        data: undefined,
+      };
+
+    const validateValue = numberTypeParamValidator(id);
+
+    if (validateValue.error)
+      return { status: 500, error: validateValue.error, date: undefined };
+
+    if (admin) {
+      const product = await makeProductUtil(Product).deleteById(id);
+
+      return {
+        status: 200,
+        error: undefined,
+        data: { massage: `${product} is successfully removed.` },
+      };
+    }
+
+    const { customer_id } = await performAuthentication(auth).validateIdParam();
+
+    if (customer_id === parseInt(id)) {
+      const product = await makeProductUtil(Product).deleteById(id);
+
+      return {
+        status: 200,
+        error: undefined,
+        data: { massage: `${product} is successfully removed.` },
+      };
+    }
 
     return {
       status: 200,
-      error: undefined,
-      data: { massage: `${product} is successfully removed.` },
+      error: "id param does not match credential id.",
+      data: undefined,
     };
   }
 }
